requirements.txt
google-generativeai
discord
httpx